requirements.txt
langgraph>=0.3.0
langgraph-checkpoint-postgres>=2.0.0
langchain-anthropic>=0.3.0
langchain-ollama>=0.3.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchainhub>=0.1.0
langsmith>=0.2.0
anthropic>=0.40.0
pydantic>=2.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
redis>=5.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
tiktoken>=0.7.0
langchain-mcp-adapters>=0.2.0
mcp>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
discord.py>=2.3.0
langfuse>=3.0.0
